# public/openapi.yaml
openapi: "3.1.0"
info:
  title: TEL & CASH — Fluxitron Custom Store Connector API
  version: "1.0.0"
  description: |
    REST API for integrating TEL & CASH e-commerce store with Fluxitron Hub.
    Manages products, prices, stock, orders, categories, and webhooks.

    **Authentication**: API Key sent in the `X-Api-Key` HTTP header.

    **Rate Limits**: 120 requests/minute per API key.
  contact:
    name: TEL & CASH
    url: https://tel-and-cash.vercel.app
servers:
  - url: https://tel-and-cash.vercel.app/api/v1
    description: Production API
security:
  - apiKey: []
paths:
...
  /stock/batch:
    post:
      operationId: updateStockBatch
      summary: Batch stock update
      description: |
        Update inventory quantities. Quantities are absolute values, not deltas.
        quantity: 0 = out of stock.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [updates]
              properties:
                updates:
                  type: array
                  items:
                    $ref: "#/components/schemas/StockUpdate"
      responses:
        "200":
          description: Batch result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchResult"
...
  /orders/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: updateOrderStatus
      summary: Update order fulfillment status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fulfillmentStatus:
                  type: string
                  enum: [unfulfilled, partial, fulfilled, cancelled]
      responses:
        "200":
          description: Status updated
  /orders/{id}/notes:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: addOrderNote
      summary: Add order note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              properties:
                note:
                  type: string
      responses:
        "201":
          description: Note added
  /orders/{id}/tracking:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: addTracking
      summary: Add tracking info
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackingInfo"
      responses:
        "200":
          description: Tracking added
  /categories:
    get:
      operationId: listCategories
      summary: List categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                parentId:
                  type: string
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key configured during store setup.
...
  schemas:
...
    TrackingInfo:
      type: object
      required: [number]
      properties:
        number:
          type: string
        company:
          type: string
        url:
          type: string
          format: uri
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        parentId:
          type: string
...
    StockUpdate:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
        variantId:
          type: string
        quantity:
          type: integer
          minimum: 0
    BatchResult:
      type: object
      required: [success, failed, errors]
      properties:
        success:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: object
            required: [id, error]
            properties:
              id:
                type: string
              error:
                type: string